requirements.txt
alembic==1.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.0.0
databricks-cli==0.16.2
dataclasses==0.6
docker==5.0.3
entrypoints==0.3
Flask==2.0.2
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mlflow==1.21.0
numpy==1.19.5
packaging==21.3
pandas==1.1.5
Pillow==8.2.0
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.6
protobuf==3.19.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.26.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
tabulate==0.8.9
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2
zipp==3.6.0
